针对log类做了一些调整

+++ src/can/IntelMatrixParse.cpp
@@ -20,7 +20,7 @@ DWORD IntelMatrixParse::decode(const BYTE* __buf, const BYTE __start_byte, const
         if (__start_byte+i>8) break;
         tmp += (__buf[__start_byte+i] << (8*i));
     }
-    return BitMask(tmp, (DWORD)__start_bit, (DWORD)__len);
+    return BitMask(tmp, __start_bit, __len);
 }
 
 DWORD IntelMatrixParse::decode(const BYTE* __buf, const BYTE __start_bit, const BYTE __len)
